GitPilot/requirements.txt
PyQt5
PyGithub
